backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
exceptiongroup==1.2.2
fastapi==0.115.12
future==1.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
junos-eznc==2.7.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
napalm==5.0.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.5.0
netutils==1.12.0
ntc_templates==7.8.0
paramiko==3.5.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyeapi==1.0.4
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
textfsm==1.1.3
transitions==0.9.2
ttp==0.9.5
ttp-templates==0.3.7
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
yamlordereddictloader==0.4.2
